Make TaskList component
