+++ client/src/components/index.js
@@ -1,1 +1,2 @@
 export { default as TaskForm } from './TaskForm'
+export { default as TaskList } from './TaskList'
